% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin2xyz.GlobColour.l3b.4km.R
\name{bin2xyz.GlobColour.l3b.4km}
\alias{bin2xyz.GlobColour.l3b.4km}
\title{Read L3b data from GlobColour project (4.6 km resolution) and return a text string in xyz format
suitable for GMT mapping tools}
\usage{
bin2xyz.GlobColour.l3b.4km(filename, prodname)
}
\arguments{
\item{filename}{is the file name string}

\item{prodname}{if the string of the product name you want to retreive (e.g. "CHL1_mean")}
}
\value{
Returns a text string in xyz format
}
\description{
Read L3b data from GlobColour project (4.6 km resolution) and return a text string in xyz format
suitable for GMT mapping tools
}
\author{
Simon Bélanger
}
