% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.l3b.timeseries.R
\name{create.l3b.timeseries}
\alias{create.l3b.timeseries}
\title{Create time series of each binned pixel found in a list of
L3b files}
\usage{
create.l3b.timeseries(latrange = c(55, 83), fnames)
}
\arguments{
\item{latrange}{is the range of latitude to consider (Default is c(55,83))}

\item{fnames}{if a vector of L3b 4km file names in netcdf format coming from Globcolour data set}
}
\value{
Return a list containing the matrix (mat) of pixels time series (column = time),
       a vector of index of the Global ISIN grid (Gindex is a unique number starting at 1 in the south pole)
}
\description{
Create time series of each binned pixel found in a list of
L3b files
}
\author{
Simon Bélanger
}
