% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.Gindex.L3b.4km.R
\name{get.Gindex.L4b.4km}
\alias{get.Gindex.L4b.4km}
\title{Convert the row/column of a pixel into an index of the 4.6 km Global ISIN grid
(see Globcolour Product User's Documentation for more details on the ISIN grid)}
\usage{
get.Gindex.L4b.4km(ixlat, ixlon)
}
\arguments{
\item{ixlat}{is the row number of the pixel}

\item{ixlon}{is the column number of the pixel}
}
\value{
The global index is return, which starts at 1 in the south pole and
finishes at 23761676 at the North pole. Note that the ISIN grid is not a regular matrix.
Each pixel of the grid have aproximately the same size (here 4.6 km x 4.6 km)
everywhere on the planet.
}
\description{
Convert the row/column of a pixel into an index of the 4.6 km Global ISIN grid
(see Globcolour Product User's Documentation for more details on the ISIN grid)
}
\author{
Simon Bélanger
}
