% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gindex2latlon.R
\name{Gindex2latlon}
\alias{Gindex2latlon}
\title{Convert Global index to lat/long}
\usage{
Gindex2latlon(Gindex)
}
\arguments{
\item{Gindex}{is the global index from the 4.6km ISIN grid}
}
\value{
Returns a data frame with the lat and long in degree
}
\description{
Convert Global index to lat/long
}
\author{
Simon Bélanger
}
