% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.GlobColour.l3b.4km.R
\name{read.GlobColour.l3b.4km}
\alias{read.GlobColour.l3b.4km}
\title{Read L3b data from GlobColour project (4.6 km resolution)}
\usage{
read.GlobColour.l3b.4km(filename, prodname)
}
\arguments{
\item{filename}{is the file name string}

\item{prodname}{if the string of the product name you want to retreive (e.g. "CHL1_mean")}
}
\value{
Return a data frame with the global index and the values of the pixels available in the file
}
\description{
Read L3b data from GlobColour project (4.6 km resolution)
}
\author{
Simon Bélanger
}
